// src/app/04_SIR.Exportaciones.Reportes/Interfaces/DatoCargaExpo.interface.ts
export interface DatoCargaExpo {
    boxid: string;            
    container: string;
    expiredate: string;         
    exportdate: string;
    finalizada: boolean;        
    grossweight: number;       
    id_Carga: number;           
    id_Pallet: number;          
    netweight: number;         
    pallet: number;            
    productcode: string;        
    productiondate: string;    
}